feat(app): move graphql interfaces to own module
move common graphql interfaces to their own separate module.
This means that all schema related code is now in a singluar location
and allows different components to easily use these interfaces.
