+++ app/src/components/search/suggestions.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import { PlayerSuggestion, PlayerSuggestionData } from "./search";
+import { PlayerSuggestion, PlayerSuggestionData } from "../../nhlapi/schema";
 
 
 /**
