+++ app/src/nhlapi/schema/index.tsx
@@ -1,0 +1,2 @@
+export * from "./player";
+export * from "./suggestion";
